+++ popout/src/main/java/org/infobip/lib/popout/backend/WalFiles.java
@@ -16,6 +16,7 @@
 
 package org.infobip.lib.popout.backend;
 
+import static java.util.Optional.ofNullable;
 import static lombok.AccessLevel.PRIVATE;
 
 import java.nio.file.Files;
@@ -23,8 +24,13 @@ import java.nio.file.Path;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Supplier;
 
+import org.infobip.lib.popout.FileQueue;
 import org.infobip.lib.popout.WalFilesConfig;
+import org.infobip.lib.popout.exception.CorruptedDataException;
 
 import io.appulse.utils.Bytes;
 import io.appulse.utils.ReadBytesUtils;
@@ -36,28 +42,43 @@ import lombok.experimental.FieldDefaults;
 import lombok.val;
 
 @FieldDefaults(level = PRIVATE, makeFinal = true)
-class WalFiles implements Iterable<WalContent> {
+class WalFiles implements Iterable<WalContent>, AutoCloseable {
 
   FilesManager files;
 
   int maxCount;
 
+  Function<CorruptedDataException, Boolean> corruptionHandler;
+
   @Builder
   WalFiles (@NonNull String queueName,
-            @NonNull Boolean restoreFromDisk,
-            @NonNull WalFilesConfig config
+            @NonNull WalFilesConfig config,
+            Boolean restoreFromDisk,
+            Function<CorruptedDataException, Boolean> corruptionHandler
   ) {
+    val restoreFromDiskValue = ofNullable(restoreFromDisk)
+        .orElse(Boolean.TRUE);
+
+    val corruptionHandlerValue = ofNullable(corruptionHandler)
+        .orElseGet(() -> new FileQueue.DefaultCorruptionHandler());
+
     files = FilesManager.builder()
         .folder(config.getFolder())
         .prefix(queueName + '-')
         .suffix(".wal")
         .build();
 
-    if (!restoreFromDisk) {
+    if (!restoreFromDiskValue) {
       files.clear();
     }
 
     maxCount = config.getMaxCount();
+    this.corruptionHandler = corruptionHandlerValue;
+  }
+
+  @Override
+  public void close () {
+    files.close();
   }
 
   void write (@NonNull Bytes buffer) {
@@ -67,17 +88,12 @@ class WalFiles implements Iterable<WalContent> {
 
   @SneakyThrows
   int pollTo (@NonNull Bytes buffer) {
-    val file = files.poll();
-    val readed = readTo(file, buffer);
-    if (readed > 0) {
-      files.remove(file);
-    }
-    return readed;
+    return readTo(buffer, files::poll, files::remove);
   }
 
   @SneakyThrows
   int peakTo (@NonNull Bytes buffer) {
-    return readTo(files.peek(), buffer);
+    return readTo(buffer, files::peek, null);
   }
 
   boolean isLimitExceeded () {
@@ -106,19 +122,36 @@ class WalFiles implements Iterable<WalContent> {
     return new WalFilesIterator();
   }
 
-  @SneakyThrows
-  private int readTo (Path file, Bytes buffer) {
-    if (file == null) {
-      return 0;
-    }
+  @SuppressWarnings("PMD.AvoidInstantiatingObjectsInLoops")
+  private int readTo (Bytes buffer, Supplier<Path> supplier, Consumer<Path> actionAfter) {
+    val writerIndex = buffer.writerIndex();
+    val readerIndex = buffer.readerIndex();
+    do {
+      val file = supplier.get();
+      try {
+        if (file == null) {
+          return 0;
+        }
 
-    val size = (int) Files.size(file);
-    if (!buffer.isWritable(size)) {
-      val newCapacity = buffer.writerIndex() + size;
-      buffer.capacity(newCapacity);
-    }
+        val size = (int) Files.size(file);
+        if (!buffer.isWritable(size)) {
+          val newCapacity = buffer.writerIndex() + size;
+          buffer.capacity(newCapacity);
+        }
 
-    return ReadBytesUtils.read(file, buffer);
+        val readed = ReadBytesUtils.read(file, buffer);
+        if (readed > 0 && actionAfter != null) {
+          actionAfter.accept(file);
+        }
+        return readed;
+      } catch (Exception ex) {
+        val exception = new CorruptedDataException(file, 0, ex);
+        corruptionHandler.apply(exception);
+        files.remove(file);
+      }
+      buffer.writerIndex(writerIndex);
+      buffer.readerIndex(readerIndex);
+    } while (true);
   }
 
   @FieldDefaults(level = PRIVATE)